+++ src/constraints/load_balance.jl
@@ -34,23 +34,24 @@ function add_elec_load_balance_constraints(m, p; _n="")
     if isempty(p.s.electric_tariff.export_bins)
         conrefs = @constraint(m, [ts in p.time_steps_with_grid],
             sum(p.production_factor[t, ts] * p.levelization_factor[t] * m[Symbol("dvRatedProduction"*_n)][t,ts] for t in p.techs.elec) +  
-            sum( m[Symbol("dvDischargeFromStorage"*_n)][b,ts] for b in p.s.storage.types ) + 
+            sum( m[Symbol("dvDischargeFromStorage"*_n)][b,ts] for b in p.s.storage.types.elec ) + 
             sum(m[Symbol("dvGridPurchase"*_n)][ts, tier] for tier in 1:p.s.electric_tariff.n_energy_tiers) ==
-            sum( sum(m[Symbol("dvProductionToStorage"*_n)][b, t, ts] for b in p.s.storage.types) 
+            sum( sum(m[Symbol("dvProductionToStorage"*_n)][b, t, ts] for b in p.s.storage.types.elec) 
                 + m[Symbol("dvCurtail"*_n)][t, ts] for t in p.techs.elec)
-            + sum(m[Symbol("dvGridToStorage"*_n)][b, ts] for b in p.s.storage.types)
+            + sum(m[Symbol("dvGridToStorage"*_n)][b, ts] for b in p.s.storage.types.elec)
+            + sum(m[Symbol("dvThermalProduction"*_n)][t, ts] / p.cop[t] for t in p.techs.cooling)
             + p.s.electric_load.loads_kw[ts]
         )
     else
         conrefs = @constraint(m, [ts in p.time_steps_with_grid],
             sum(p.production_factor[t, ts] * p.levelization_factor[t] * m[Symbol("dvRatedProduction"*_n)][t,ts] for t in p.techs.elec) +  
-            sum( m[Symbol("dvDischargeFromStorage"*_n)][b,ts] for b in p.s.storage.types ) + 
+            sum( m[Symbol("dvDischargeFromStorage"*_n)][b,ts] for b in p.s.storage.types.elec ) + 
             sum(m[Symbol("dvGridPurchase"*_n)][ts, tier] for tier in 1:p.s.electric_tariff.n_energy_tiers) ==
-            sum(  sum(m[Symbol("dvProductionToStorage"*_n)][b, t, ts] for b in p.s.storage.types) 
+            sum(  sum(m[Symbol("dvProductionToStorage"*_n)][b, t, ts] for b in p.s.storage.types.elec) 
                 + sum(m[Symbol("dvProductionToGrid"*_n)][t, u, ts] for u in p.export_bins_by_tech[t]) 
                 + m[Symbol("dvCurtail"*_n)][t, ts] 
             for t in p.techs.elec)
-            + sum(m[Symbol("dvGridToStorage"*_n)][b, ts] for b in p.s.storage.types)
+            + sum(m[Symbol("dvGridToStorage"*_n)][b, ts] for b in p.s.storage.types.elec)
             + p.s.electric_load.loads_kw[ts]
         )
     end
@@ -62,8 +63,8 @@ function add_elec_load_balance_constraints(m, p; _n="")
 	##Constraint (8b): Electrical Load Balancing without Grid
 	@constraint(m, [ts in p.time_steps_without_grid],
 		sum(p.production_factor[t,ts] * p.levelization_factor[t] * m[Symbol("dvRatedProduction"*_n)][t,ts] for t in p.techs.elec) +  
-		sum( m[Symbol("dvDischargeFromStorage"*_n)][b,ts] for b in p.s.storage.types )  ==
-        sum( sum(m[Symbol("dvProductionToStorage"*_n)][b, t, ts] for b in p.s.storage.types) + 
+		sum( m[Symbol("dvDischargeFromStorage"*_n)][b,ts] for b in p.s.storage.types.elec )  ==
+        sum( sum(m[Symbol("dvProductionToStorage"*_n)][b, t, ts] for b in p.s.storage.types.elec) + 
         m[Symbol("dvCurtail"*_n)][t, ts] for t in p.techs.elec) +
         p.s.electric_load.critical_loads_kw[ts]
 	)
@@ -74,13 +75,13 @@ function add_production_constraints(m, p; _n="")
 	# Constraint (4d): Electrical production sent to storage or export must be less than technology's rated production
     if isempty(p.s.electric_tariff.export_bins)
         @constraint(m, [t in p.techs.elec, ts in p.time_steps_with_grid],
-            sum(m[Symbol("dvProductionToStorage"*_n)][b, t, ts] for b in p.s.storage.types)  
+            sum(m[Symbol("dvProductionToStorage"*_n)][b, t, ts] for b in p.s.storage.types.elec)  
           + m[Symbol("dvCurtail"*_n)][t, ts]
          <= p.production_factor[t, ts] * p.levelization_factor[t] * m[Symbol("dvRatedProduction"*_n)][t, ts]
         )
     else
         @constraint(m, [t in p.techs.elec, ts in p.time_steps_with_grid],
-            sum(m[Symbol("dvProductionToStorage"*_n)][b, t, ts] for b in p.s.storage.types)  
+            sum(m[Symbol("dvProductionToStorage"*_n)][b, t, ts] for b in p.s.storage.types.elec)  
           + m[Symbol("dvCurtail"*_n)][t, ts]
           + sum(m[Symbol("dvProductionToGrid"*_n)][t, u, ts] for u in p.export_bins_by_tech[t])
          <= p.production_factor[t, ts] * p.levelization_factor[t] * m[Symbol("dvRatedProduction"*_n)][t, ts]
@@ -89,7 +90,7 @@ function add_production_constraints(m, p; _n="")
 
 	# Constraint (4e): Electrical production sent to storage or grid must be less than technology's rated production - no grid
 	@constraint(m, [t in p.techs.elec, ts in p.time_steps_without_grid],
-		sum(m[Symbol("dvProductionToStorage"*_n)][b, t, ts] for b in p.s.storage.types)  <= 
+		sum(m[Symbol("dvProductionToStorage"*_n)][b, t, ts] for b in p.s.storage.types.elec)  <= 
 		p.production_factor[t, ts] * p.levelization_factor[t] * m[Symbol("dvRatedProduction"*_n)][t, ts]
 	)
 
@@ -97,21 +98,20 @@ end
 
 
 function add_thermal_load_constraints(m, p; _n="")
-	### Constraint set (5) - hot and cold thermal loads
 
-	##Constraint (5a): Cold thermal loads
-	# if !isempty(p.CoolingTechs)
-	# 	@constraint(m, ColdThermalLoadCon[ts in p.time_steps],
-	# 			sum(p.production_factor[t,ts] * m[Symbol("dvThermalProduction"*_n)][t,ts] for t in p.CoolingTechs) +
-	# 			sum(m[:dvDischargeFromStorage][b,ts] for b in p.ColdTES) ==
-	# 			p.CoolingLoad[ts] -
-	# 			sum(p.GHPCoolingThermalServed[g,ts] * m[:binGHP][g] for g in p.GHPOptions) +
-	# 			sum(m[:dvProductionToStorage][b,t,ts] for b in p.ColdTES, t in p.CoolingTechs)
-	# 	)
-	# end
+    m[Symbol("binFlexHVAC"*_n)] = 0
+    m[Symbol("dvTemperature"*_n)] = 0
+    m[Symbol("dvComfortLimitViolationCost"*_n)] = 0
+
+    if !isnothing(p.s.flexible_hvac)
+        #= FlexibleHVAC does not require equality constraints for thermal loads. The thermal loads
+        are instead a function of the energy required to keep the space temperature within the 
+        comfort limits.
+        =#
+        add_flexible_hvac_constraints(m, p, _n=_n) 
 
 	##Constraint (5b): Hot thermal loads
-	if !isempty(p.techs.heating)
+    elseif !isempty(p.techs.heating)
         
         # if !isempty(p.SteamTurbineTechs)
         #     @constraint(m, HotThermalLoadCon[ts in p.time_steps],
@@ -119,11 +119,11 @@ function add_thermal_load_constraints(m, p; _n="")
         #             # sum(m[Symbol("dvThermalProduction"*_n)][t,ts] for t in p.SteamTurbineTechs) +
         #             sum(p.production_factor[t,ts] * (m[Symbol("dvThermalProduction"*_n)][t,ts] - m[:dvThermalToSteamTurbine][t,ts]) for t in p.techs.boiler)
         #             # + sum(p.GHPHeatingThermalServed[g,ts] * m[:binGHP][g] for g in p.GHPOptions)
-        #             # + sum(m[:dvDischargeFromStorage][b,ts] for b in p.HotTES) 
+        #             # + sum(m[:dvDischargeFromStorage][b,ts] for b in p.s.storage.types.hot) 
         #             ==
         #             p.HeatingLoad[ts] * p.s.existing_boiler.efficiency
         #             # + sum(m[:dvProductionToWaste][t,ts] for t in p.CHPTechs) +
-        #             # sum(m[:dvProductionToStorage][b,t,ts] for b in p.HotTES, t in p.techs.heating) +
+        #             # sum(m[:dvProductionToStorage][b,t,ts] for b in p.s.storage.types.hot, t in p.techs.heating) +
         #             # sum(m[Symbol("dvThermalProduction"*_n)][t,ts] for t in p.AbsorptionChillers) / p.AbsorptionChillerCOP
         #     )
         # else
@@ -131,14 +131,32 @@ function add_thermal_load_constraints(m, p; _n="")
                     sum(m[Symbol("dvThermalProduction"*_n)][t,ts] for t in p.techs.chp) +
                     sum(m[Symbol("dvThermalProduction"*_n)][t, ts] for t in p.techs.boiler)
                     # TODO do all thermal techs have production_factor = 1 ? get rid of it if so
-                    # + sum(m[:dvDischargeFromStorage][b,ts] for b in p.HotTES)
+                    + sum(m[:dvDischargeFromStorage][b,ts] for b in p.s.storage.types.hot)
                     # + sum(p.GHPHeatingThermalServed[g,ts] * m[:binGHP][g] for g in p.GHPOptions)
                     ==
                     (p.s.dhw_load.loads_kw[ts] + p.s.space_heating_load.loads_kw[ts])
-                    + sum(m[Symbol("dvProductionToWaste"*_n)][t,ts] for t in p.techs.chp) #+
-                    # sum(m[:dvProductionToStorage][b,t,ts] for b in p.HotTES, t in p.techs.heating)  +
+                    + sum(m[Symbol("dvProductionToWaste"*_n)][t,ts] for t in p.techs.chp) +
+                    sum(m[:dvProductionToStorage][b,t,ts] for b in p.s.storage.types.hot, t in p.techs.heating)  #+
                     # sum(m[Symbol("dvThermalProduction"*_n)][t,ts] for t in p.AbsorptionChillers) / p.AbsorptionChillerCOP
             )
         # end
-	end
+
+	##Constraint (5a): Cold thermal loads
+	# if !isempty(p.CoolingTechs)
+	# 	@constraint(m, ColdThermalLoadCon[ts in p.time_steps],
+	# 			sum(p.production_factor[t,ts] * m[Symbol("dvThermalProduction"*_n)][t,ts] for t in p.CoolingTechs) +
+	# 			sum(m[:dvDischargeFromStorage][b,ts] for b in p.s.storage.types.cold) ==
+	# 			p.CoolingLoad[ts] -
+	# 			sum(p.GHPCoolingThermalServed[g,ts] * m[:binGHP][g] for g in p.GHPOptions) +
+	# 			sum(m[:dvProductionToStorage][b,t,ts] for b in p.s.storage.types.cold, t in p.CoolingTechs)
+	# 	)
+	# end
+
+    # TODO do we need production_factor for chillers?
+    elseif !isempty(p.techs.cooling)
+        @constraint(m, [ts in p.time_steps],
+            sum(m[Symbol("dvThermalProduction"*_n)][t, ts] for t in p.techs.cooling) ==
+            p.s.cooling_load.loads_kw_thermal[ts]
+        )
+    end
 end