+++ src/core/reopt.jl
@@ -148,7 +148,7 @@ function build_reopt!(m::JuMP.AbstractModel, p::REoptInputs)
 			fix(m[:dvGridPurchase][ts, tier] , 0.0, force=true)
 		end
 
-		for t in p.s.storage.types
+		for t in p.s.storage.types.elec
 			fix(m[:dvGridToStorage][t, ts], 0.0, force=true)
 		end
 
@@ -159,22 +159,33 @@ function build_reopt!(m::JuMP.AbstractModel, p::REoptInputs)
         end
 	end
 
-	for b in p.s.storage.types
-		if p.s.storage.max_kw[b] == 0 || p.s.storage.max_kwh[b] == 0
+	for b in p.s.storage.types.all
+		if p.s.storage.attr[b].max_kw == 0 || p.s.storage.attr[b].max_kwh == 0
 			@constraint(m, [ts in p.time_steps], m[:dvStoredEnergy][b, ts] == 0)
 			@constraint(m, m[:dvStorageEnergy][b] == 0)
-			@constraint(m, m[:dvStoragePower][b] == 0)
-			@constraint(m, [t in p.techs.elec, ts in p.time_steps_with_grid],
-						m[:dvProductionToStorage][b, t, ts] == 0)
 			@constraint(m, [ts in p.time_steps], m[:dvDischargeFromStorage][b, ts] == 0)
-			@constraint(m, [ts in p.time_steps], m[:dvGridToStorage][b, ts] == 0)
+			if b in p.s.storage.types.elec
+				@constraint(m, m[:dvStoragePower][b] == 0)
+				@constraint(m, [ts in p.time_steps], m[:dvGridToStorage][b, ts] == 0)
+				@constraint(m, [t in p.techs.elec, ts in p.time_steps_with_grid],
+						m[:dvProductionToStorage][b, t, ts] == 0)
+			end
 		else
 			add_storage_size_constraints(m, p, b)
-			add_storage_dispatch_constraints(m, p, b)
+			add_general_storage_dispatch_constraints(m, p, b)
+			if b in p.s.storage.types.elec
+				add_elec_storage_dispatch_constraints(m, p, b)
+			elseif b in p.s.storage.types.hot
+				add_hot_thermal_storage_dispatch_constraints(m, p, b)
+			elseif b in p.s.storage.types.cold
+				add_cold_thermal_storage_dispatch_constraints(m, p, b)
+			else
+				@error("Invalid storage does not fall in a thermal or electrical set")
+			end
 		end
 	end
 
-	if any(max_kw->max_kw > 0, (p.s.storage.max_kw[b] for b in p.s.storage.types))
+	if any(max_kw->max_kw > 0, (p.s.storage.attr[b].max_kw for b in p.s.storage.types.elec))
 		add_storage_sum_constraints(m, p)
 	end
 
@@ -185,6 +196,7 @@ function build_reopt!(m::JuMP.AbstractModel, p::REoptInputs)
     m[:TotalPerUnitHourOMCosts] = 0.0
     m[:TotalFuelCosts] = 0.0
     m[:TotalProductionIncentive] = 0
+	m[:dvComfortLimitViolationCost] = 0.0
 	m[:TotalCHPStandbyCharges] = 0
 
 	if !isempty(p.techs.all)
@@ -275,10 +287,10 @@ function build_reopt!(m::JuMP.AbstractModel, p::REoptInputs)
         end
     end
 	
-	@expression(m, TotalStorageCapCosts, p.third_party_factor *
-		sum(  p.s.storage.installed_cost_per_kw[b] * m[:dvStoragePower][b]
-			+ p.s.storage.installed_cost_per_kwh[b] * m[:dvStorageEnergy][b] for b in p.s.storage.types )
-	)
+	@expression(m, TotalStorageCapCosts, p.third_party_factor * (
+		sum( p.s.storage.attr[b].net_present_cost_per_kw * m[:dvStoragePower][b] for b in p.s.storage.types.elec) + 
+		sum( p.s.storage.attr[b].net_present_cost_per_kwh * m[:dvStorageEnergy][b] for b in p.s.storage.types.all )
+	))
 	
 	@expression(m, TotalPerUnitSizeOMCosts, p.third_party_factor * p.pwf_om *
 		sum( p.om_cost_per_kw[t] * m[:dvSize][t] for t in p.techs.all )
@@ -290,7 +302,11 @@ function build_reopt!(m::JuMP.AbstractModel, p::REoptInputs)
 		add_dv_UnservedLoad_constraints(m,p)
 		add_outage_cost_constraints(m,p)
 		add_MG_production_constraints(m,p)
-		add_MG_storage_dispatch_constraints(m,p)
+		if !isempty(p.s.storage.types.elec)
+			add_MG_storage_dispatch_constraints(m,p)
+		else
+			fix_MG_storage_variables(m,p)
+		end
 		add_cannot_have_MG_with_only_PVwind_constraints(m,p)
 		add_MG_size_constraints(m,p)
 		
@@ -331,17 +347,20 @@ function build_reopt!(m::JuMP.AbstractModel, p::REoptInputs)
 		m[:TotalElecBill] * (1 - p.s.financial.offtaker_tax_pct) -
 
         # Subtract Incentives, which are taxable
-		m[:TotalProductionIncentive] * (1 - p.s.financial.owner_tax_pct)
+		m[:TotalProductionIncentive] * (1 - p.s.financial.owner_tax_pct) +
+
+		m[:dvComfortLimitViolationCost]
 	);
 	if !isempty(p.s.electric_utility.outage_durations)
 		add_to_expression!(Costs, m[:ExpectedOutageCost] + m[:mgTotalTechUpgradeCost] + m[:dvMGStorageUpgradeCost] + m[:ExpectedMGFuelCost])
 	end
 
-	if !p.s.settings.add_soc_incentive
-		@objective(m, Min, m[:Costs])
-	else  # Keep SOC high
-		@objective(m, Min, m[:Costs] - sum(m[:dvStoredEnergy][:elec, ts] for ts in p.time_steps) /
-									   (8760. / p.hours_per_timestep)
+	@objective(m, Min, m[:Costs])
+	
+	if !(isempty(p.s.storage.types.elec)) && p.s.settings.add_soc_incentive # Keep SOC high
+		@objective(m, Min, m[:Costs] - 
+		sum(m[:dvStoredEnergy][b, ts] for b in p.s.storage.types.elec, ts in p.time_steps) /
+			(8760. / p.hours_per_timestep)
 		)
 	end
     
@@ -394,12 +413,12 @@ function add_variables!(m::JuMP.AbstractModel, p::REoptInputs)
 		dvGridPurchase[p.time_steps, 1:p.s.electric_tariff.n_energy_tiers] >= 0  # Power from grid dispatched to meet electrical load [kW]
 		dvRatedProduction[p.techs.all, p.time_steps] >= 0  # Rated production of technology t [kW]
 		dvCurtail[p.techs.all, p.time_steps] >= 0  # [kW]
-		dvProductionToStorage[p.s.storage.types, p.techs.all, p.time_steps] >= 0  # Power from technology t used to charge storage system b [kW]
-		dvDischargeFromStorage[p.s.storage.types, p.time_steps] >= 0 # Power discharged from storage system b [kW]
-		dvGridToStorage[p.s.storage.types, p.time_steps] >= 0 # Electrical power delivered to storage by the grid [kW]
-		dvStoredEnergy[p.s.storage.types, 0:p.time_steps[end]] >= 0  # State of charge of storage system b
-		dvStoragePower[p.s.storage.types] >= 0   # Power capacity of storage system b [kW]
-		dvStorageEnergy[p.s.storage.types] >= 0   # Energy capacity of storage system b [kWh]
+		dvProductionToStorage[p.s.storage.types.all, p.techs.all, p.time_steps] >= 0  # Power from technology t used to charge storage system b [kW]
+		dvDischargeFromStorage[p.s.storage.types.all, p.time_steps] >= 0 # Power discharged from storage system b [kW]
+		dvGridToStorage[p.s.storage.types.elec, p.time_steps] >= 0 # Electrical power delivered to storage by the grid [kW]
+		dvStoredEnergy[p.s.storage.types.all, 0:p.time_steps[end]] >= 0  # State of charge of storage system b
+		dvStoragePower[p.s.storage.types.all] >= 0   # Power capacity of storage system b [kW]
+		dvStorageEnergy[p.s.storage.types.all] >= 0   # Energy capacity of storage system b [kWh]
 		dvPeakDemandTOU[p.ratchets, 1:p.s.electric_tariff.n_tou_demand_tiers] >= 0  # Peak electrical power demand during ratchet r [kW]
 		dvPeakDemandMonth[p.months, 1:p.s.electric_tariff.n_monthly_demand_tiers] >= 0  # Peak electrical power demand during month m [kW]
 		MinChargeAdder >= 0