+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # REopt Changelog
 
+## develop
+- fix bugs for time_steps_per_hour != 1
+- add FlexibleHVAC model (still testing)
+- start thermal energy storage modeling
+- refactor `Storage` as `ElectricStorage`
+- add `ExistingBoiler` and `ExistingChiller`
+
+## v0.12.4
+- rm "Lite" from docs
+- prioritize `urdb_response` over `urdb_label` in `ElectricTariff`
+
 ## v0.12.3
 - add utils for PVwatts: `get_ambient_temperature` and `get_pvwatts_prodfactor`